Fix 32/64 endian for all services from a previous missed commit.
git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@1094 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ exec/ipc.c
@@ -158,7 +158,7 @@ static int response_init_send_response (
 	res_lib_response_init.header.size = sizeof (mar_res_lib_response_init_t);
 	res_lib_response_init.header.id = MESSAGE_RES_INIT;
 	res_lib_response_init.header.error = error;
-	res_lib_response_init.conn_info = conn_info;
+	res_lib_response_init.conn_info = (unsigned long long)conn_info;
 
 	openais_conn_send_response (
 		conn_info,
